size_t is not an unsigned int on all systems
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@17746 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ plugins/nagware/nagware.cpp
@@ -118,9 +118,9 @@ void updatePlayerNextEvent (int playerID, double now){
 
 void sendNagMessage (int who, std::string *msg ){
   std::string fullMsg = *msg + Config.msgSuffix;
-  unsigned int idx=0, x;
+  size_t idx=0, x;
 
-  while ((x = (unsigned int)fullMsg.find("\\n", idx)) != std::string::npos){
+  while ((x = fullMsg.find("\\n", idx)) != std::string::npos){
     bz_sendTextMessage(BZ_SERVER, who, fullMsg.substr(idx, x-idx).c_str());
     idx = x+2;
   }
